requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
cytoolz==0.12.3
dask==2024.5.1
dask-expr==1.1.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
fsspec==2024.5.0
hexbytes==0.3.1
idna==3.7
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
locket==1.0.0
lru-dict==1.2.0
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parsimonious==0.10.0
partd==1.4.2
protobuf==5.26.1
pyarrow==16.1.0
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
web3==6.19.0
websockets==12.0
yarl==1.9.4
